requirements/base.txt
certifi==2019.11.28
celery==4.4.0
Django==3.0.2
django-anymail==7.0.0
django-celery-beat==1.5.0
django-celerybeat-status==0.0.10
django-cors-headers==3.2.0
django-countries==5.5
django-ckeditor==5.8.0
django-environ==0.4.5
django-filter==2.2.0
django-modeltranslation==0.14.1
django-money==1.0
django-mptt==0.10.0
django-mptt-admin==1.0.0
django-optimized-image==0.2.0
django-rest-swagger==2.2.0
django-split-settings==1.0.0
djangorestframework==3.11.0
extensions==0.4
idna==2.8
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.11.1
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sentry-sdk==0.13.5
social-auth-app-django==3.1.0
social-auth-core==3.2.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
django-ckeditor==5.8.0
django-registration==3.0.1
sendgrid==6.1.0
django-grappelli==2.13.3
django-redis==4.11.0
drf-yasg==1.17.0
